Use semantically appropriate integer types throughout
- i32 arguments and fields that are semantically non-negative are changed
  to u32, u8, u16, or usize as appropriate (volumes, velocities, sample
  counts, channel counts, sample rates, beat parameters, etc.)
- Remove unnecessary `as` casts made redundant by the type changes
- VomitPreparation::flags: u32 → u8; VomitPrepFlags constants likewise
- PxtoneService: dst_ch_num u8, dst_sps u32, moo_smp_* / moo_fade_* u32,
  moo_bt_clock u16, moo_bt_num u8
- Remove unused pub(crate) helpers (count_v5, count_x4x, set_meas_num,
  set_beat_clock, get_this_clock, table_size, get_sec, buf_size, read_wav,
  convert_volume, get_sample_i16_at) and unused constants
- Fix i16 overflow in noise_builder Saw6 wave segment computation
- Clamp start_pos_sample to 0 when negative (matches C++ intent)

+++ src/effect/delay.rs
@@ -60,7 +60,7 @@ impl Delay {
   }
 
   /// Prepares before playback: allocates the delay buffer
-  pub(crate) fn tone_ready(&mut self, beat_num: i32, beat_tempo: f32, sps: i32) {
+  pub(crate) fn tone_ready(&mut self, beat_num: u8, beat_tempo: f32, sps: u32) {
     self.smp_num = 0;
     self.bufs[0].clear();
     self.bufs[1].clear();
